Add files via upload
Given an array arr[] of size N-1 with integers in the range of [1, N], the task is to find the missing number from the first N integers
Find the sum of the numbers in the range [1, N] using the formula N * (N+1)/2. Now find the sum of all the elements in the array and subtract it from the sum of the first N natural numbers. This will give the value of the missing element.